Some adjustments to work better on mobile

+++ snake_chef/lib/game/stage.dart
@@ -15,7 +15,7 @@ class Stage {
   factory Stage.fromJson(Map<String, dynamic> json) => Stage(
         time: json["time"],
         recipes: json["recipes"].map((recipe) => Recipe.fromJson(recipe)).toList().cast<Recipe>(),
-        obstacles: json["obstacles"].map((obstacle) => Position(obstacle["x"], obstacle["y"])).toList().cast<Position>(),
+        obstacles: json["obstacles"].map((obstacle) => Position(obstacle["x"].toDouble(), obstacle["y"].toDouble())).toList().cast<Position>(),
         initialX: json["initialX"],
         initialY: json["initialY"],
       );
